refactor(reconciler): give the affinity decode its own file again
Reverses the file move from ef3dd4f, because the argument behind it no longer
applies. That commit folded affinity.go into role_group_handler.go on the
grounds that it separated two siblings — mergeAffinity from mergeResources —
which made reading MergeRoleGroupConfig a two-file exercise.

mergeAffinity is gone. Affinity is replaced wholesale, so the merge is three
lines inline in MergeRoleGroupConfig and there is no helper left to keep beside
mergeResources. What remains is DecodeAffinity: how the framework READS one
schema-free CRD field, whose only caller is in base_role_group_handler.go and
which has nothing to do with the config fold.

Keeping it in role_group_handler.go made that file hold resource structs, build
contexts, provider interfaces, logging renderers, the config merge, the handler
interface AND an unrelated decode helper. A 71-line affinity.go says exactly
what it is, and it re-pairs with affinity_test.go, which was left orphaned.

Two things also cleaned up in the move: the file-level prose sat before the
package clause, where Go would read it as a second package doc comment, so it is
folded into DecodeAffinity's own doc; and a comment still referred to
affinityMembers, deleted in 364ba45.

No behaviour change.
